Safari-611.1.10/LayoutTests/platform/glib/fast/canvas/webgl/vertexAttribPointer-with-bad-offset-expected.txt

CONSOLE MESSAGE: WebGL: INVALID_VALUE: vertexAttribPointer: bad offset
CONSOLE MESSAGE: WebGL: INVALID_OPERATION: drawArrays: attempt to access out of bounds arrays
PASS: vertexAttribPointer should have an error.
